cru_cy column names

A string list containing column names of the raw CEDA cru_cy data.

cru_cy_colnames

Format

A string list with 18 entries.

YEAR

The string "YEAR", representing the year represnted in the data.

JAN

The string "JAN", representing a time period consisting of the month January.

FEB

The string "FEB", representing a time period consisting of the month February.

MAR

The string "MAR", representing a time period consisting of the month March.

APR

The string "APR", representing a time period consisting of the month April.

MAY

The string "MAY", representing a time period consisting of the month May.

JUN

The string "JUN", representing a time period consisting of the month June.

JUL

The string "JUL", representing a time period consisting of the month July.

AUG

The string "AUG", representing a time period consisting of the month August.

SEP

The string "SEP", representing a time period consisting of the month September.

OCT

The string "OCT", representing a time period consisting of the month October.

NOV

The string "NOV", representing a time period consisting of the month November.

DEC

The string "DEC", representing a time period consisting of the month December.

MAM

The string "MAM", representing a time period consisting of the months of March, April, and May.

JJA

The string "JJA", representing a time period consisting of the months of June, July, and August

SON

The string "SON", representing a time period consisting of the months of September, October, and November.

DJF

The string "DJF", representing a time period consisting of the months of December, Janaury, and February.

ANN

The string "ANN", representing a time period consisting of the full year.

Examples

cru_cy_colnames

#> $YEAR
#> [1] "YEAR"
#> 
#> $JAN
#> [1] "JAN"
#> 
#> $FEB
#> [1] "FEB"
#> 
#> $MAR
#> [1] "MAR"
#> 
#> $APR
#> [1] "APR"
#> 
#> $MAY
#> [1] "MAY"
#> 
#> $JUN
#> [1] "JUN"
#> 
#> $JUL
#> [1] "JUL"
#> 
#> $AUG
#> [1] "AUG"
#> 
#> $SEP
#> [1] "SEP"
#> 
#> $OCT
#> [1] "OCT"
#> 
#> $NOV
#> [1] "NOV"
#> 
#> $DEC
#> [1] "DEC"
#> 
#> $MAM
#> [1] "MAM"
#> 
#> $JJA
#> [1] "JJA"
#> 
#> $SON
#> [1] "SON"
#> 
#> $DJF
#> [1] "DJF"
#> 
#> $ANN
#> [1] "ANN"
#>